requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
sqlalchemy==2.0.23
pydantic==2.5.0
openai>=1.30.0
beautifulsoup4==4.12.2
requests==2.31.0
python-multipart==0.0.6
python-dotenv==1.0.0
pandas==2.0.3
aiofiles==23.2.1
httpx==0.25.0
lxml==4.9.3
playwright==1.40.0
firecrawl-py==3.3.2
